Se creo otro componente donde se aparta el formulario de registro y invoca en invoiceapp, se agrea boton de ocular y mostrar formulario de registro y por ultimo se agrega la accion de eliminar registros de la lista de productos
