+++ 4-invoice-app/src/InvoiceApp.jsx
@@ -5,6 +5,7 @@ import { EmpresaView } from "./components/EmpresaView";
 import { DetailView } from "./components/InvoiceDetailView";
 import { ItemView } from "./components/ItemView";
 import { TotalView } from "./components/TotalView";
+import { FormItemsView } from "./components/FormItemsView";
 
 
 const invoiceInitial = {
@@ -30,6 +31,9 @@ const invoiceInitial = {
 
 export const InvoiceApp = () => {
 
+    //accion para boton de ocultar y mostrar form
+    const [activeForm, setActiveForm] = useState(false);
+
     const [total, setTotal] = useState(0);
     //usa para el contador del id iniciando en 4 porque ya se tiene 3 registros
     const [counter, setCounter] = useState(4);
@@ -38,15 +42,11 @@ export const InvoiceApp = () => {
     //para guardar los items
     const [items, setItems] = useState([]);
 
-    const [formItemsState,setFormItemsState] = useState({
-        product: '',
-        price:'',
-        quantity:'',
-    });
+ 
     //factura desestructuracion
     const {id,name, client, company} = invoice;
 
-    const {product, price, quantity} = formItemsState;
+    
 
     useEffect(() =>{
         const data = getInvoice();
@@ -55,13 +55,7 @@ export const InvoiceApp = () => {
         setItems(data.items)
     }, []);
 
-    useEffect(() =>{
-       //console.log('Cambio de precio')
-    } ,[price]);
 
-    useEffect(() =>{
-        //console.log('Cambio de formItemsState')
-    } ,[formItemsState]);
 
     useEffect(() =>{
         //console.log('El counter cambio')
@@ -77,15 +71,7 @@ export const InvoiceApp = () => {
     const [priceValue, setPriceValue] = useState('');
     const [quantityValue, setQuantityValue] = useState('');*/
 
-    //se mejora para tomar los campos en una sola funcion en un solo estado
-    const onInvoiceChange = ({target: {name, value}}) => {
-        //console.log(name);
-        //console.log(value);
-        setFormItemsState({
-            ...formItemsState,
-            [name]:value
-        })
-    };
+  
     /*const onProductChange = ({target}) => {
         console.log(target.value);
         setProductValue(target.value);
@@ -100,20 +86,7 @@ export const InvoiceApp = () => {
         setQuantityValue(event.target.value);
     };*/
     
-    const onInvoiceItemSubmit = (event) => {
-        event.preventDefault();
-        //se realizan las validaciones del formulario
-        if(product.trim().length <= 1) return;
-        if(price.trim().length <= 1) {
-            alert('El campo de ser de 2 digito')
-            return
-        };
-        if(isNaN(price.trim())) {
-            alert('No es un numero '+price)
-            return;
-        }
-        if(quantity.trim().length < 1) return;
-        if(isNaN(quantity.trim())) return;
+    const handlerAddItems = ({product,price,quantity}) => {
         
         setItems([...items, {
             id:counter,
@@ -123,14 +96,19 @@ export const InvoiceApp = () => {
             //convertir entero con parseInt
             quantity:parseInt(quantity.trim(), 10)
         }]);
-        setFormItemsState({
-            product: '',
-            price: '',
-            quantity:'',
-        });
+
         setCounter(counter+1);
 
     };
+     //accion para boton de eliminar registros o items
+    const handlerDeleteItem = (id) =>{
+        setItems(items.filter(item => item.id != id))
+    }
+
+     //accion para boton de ocultar y mostrar form
+    const onActiveForm = () =>{
+        setActiveForm(!activeForm);
+    }
     return (
         <>
             <div className="container my-3">
@@ -148,47 +126,21 @@ export const InvoiceApp = () => {
                                 <EmpresaView title="Datos de la empresa" company={company}/>
                             </div>
                         </div>
-                        <ItemView title="Productos de la empresa" items={items} />
+                        <ItemView title="Productos de la empresa" items={items} handlerDeleteItem={ id => handlerDeleteItem(id) } />
                     </div>
                     <div className="card-footer">
                         <TotalView title= "Total de Compra:" total={total}/>
                     </div>
                 </div>
-
-                <form className="align-items-center card my-3" onSubmit={ onInvoiceItemSubmit }>
-                    <h3>Nuevo Item</h3>
-                    <div className="row my-3">
-                        <div className="col-6">
-                            <input 
-                            className="form-control" 
-                            type="text" 
-                            name="product" 
-                            value={product}
-                            placeholder="Producto" 
-                            onChange={ onInvoiceChange } />
-                        </div>
-                        <div className="col-3">
-                            <input 
-                            className="form-control" 
-                            type="text" 
-                            name="price" 
-                            value={price}
-                            placeholder="$Preio" 
-                            onChange={onInvoiceChange} />
-                        </div>
-                        <div className="col-3">
-                            <input className="form-control"
-                            type="number" 
-                            name="quantity" 
-                            value={quantity}
-                            placeholder="Cantidad" 
-                            onChange={ event => onInvoiceChange(event)} /> 
-                        </div>
-                    </div>
-                    <div className="my-3">
-                        <button type="sumit" className="btn btn-primary">Guardar</button>
-                    </div>
-                </form>
+                {/*Boton para ocultar el formulario y mostrar*/}
+                <button className="btn btn-secondary mt-3"
+                onClick={ onActiveForm}>{  !activeForm ? 'Agregar Item' : 'Ocultar Fromulario'}</button>
+                {/* puede ser asi */}
+                { /* {!activeForm ? '' : <FormItemsView handler={handlerAddItems}/>}*/ }
+                {/* o de esta forma*/}
+                { !activeForm || <FormItemsView handler={handlerAddItems}/> }
+                
+                
             </div>
         </>
     )
