+++ 4-invoice-app/src/components/RowItem.jsx
@@ -1,11 +1,12 @@
 import PropTypes from 'prop-types'
-export const RowItem = ({product, price, quantity}) =>{
+export const RowItem = ({id,product, price, quantity, handlerDeleteItem}) =>{
     return(
         <>
             <tr>
                 <td>{ product }</td>
                 <td>{ price }</td>
                 <td className='text-center'>{ quantity }</td> 
+                <td><button className='btn btn-danger' onClick={() => handlerDeleteItem(id)}>Eliminar</button></td>
             </tr>
         </>
     )
